+++ public/js/controllers/DashboardCtrl.js
@@ -23,6 +23,10 @@
 
     // remove reservation by _id
     $scope.remove = function (reservationID) {
+        // Confirmation dialog
+        if (!confirm("Opravdu chcete smazat rezervaci?"))
+            return;
+
         ReservationService.delete(reservationID).success(function (data) {
             if (data.isValid) {
                 loadReservations();
